fix: changed INT to FLOAT

+++ src/Hy3Layout.cpp
@@ -1540,7 +1540,7 @@ void Hy3Layout::applyNodeDataToWindow(Hy3Node* node, bool no_animation) {
 
 		if (g_pCompositor->isWorkspaceSpecial(window->m_iWorkspaceID)) {
 			// adjust for special workspaces
-			static const auto scalefactor = ConfigValue<Hyprlang::INT>("plugin:hy3:special_scale_factor");  
+			static const auto scalefactor = ConfigValue<Hyprlang::FLOAT>("plugin:hy3:special_scale_factor");  
 			calcPos = calcPos + (calcSize - calcSize * *scalefactor) / 2.f;
 			calcSize = calcSize * *scalefactor;
 		}
